Split SpecParameterParsingTest for non-osgi and osgi
